pyproject.toml
[build-system]
requires = ["setuptools>=77.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "malwoverview"
version = "8.0"
description = "Malwoverview is a first response tool for threat hunting."
readme = "README.md"
authors = [{name = "Alexandre Borges", email = "reverseexploit@proton.me"}]
license = "GPL-3.0-or-later"
requires-python = ">=3.8"
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pefile",
    "colorama",
    "python-magic>=0.4.25; platform_system == 'Linux' or platform_system == 'Darwin'",
    "python-magic-bin>=0.4.14; platform_system == 'Windows'",
    "simplejson",
    "requests",
    "urllib3>=2.6.3",
    "validators",
    "geocoder",
    "polyswarm-api",
    "tqdm>=4.64.0",
    "python-whois>=0.9.4",
    "ipwhois>=1.2.0",
]

[project.optional-dependencies]
yara = ["yara-python>=4.2.0"]
pdf = ["weasyprint>=57.0", "PyPDF2>=3.0.0"]
tui = ["textual>=0.40.0", "pyperclip>=1.8.0"]
all = ["yara-python>=4.2.0", "weasyprint>=57.0", "PyPDF2>=3.0.0", "textual>=0.40.0", "pyperclip>=1.8.0"]

[project.urls]
Homepage = "https://github.com/alexandreborges/malwoverview"
Repository = "https://github.com/alexandreborges/malwoverview"

[project.scripts]
malwoverview = "malwoverview.malwoverview:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["malwoverview", "malwoverview.*"]
